pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autonovelclaw"
version = "0.1.0"
description = "From Idea to Published Novel — Fully Autonomous. Tell me your idea. Pick a storyline. I'll write, review, enhance, and publish it."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "AutoNovelClaw Contributors"}
]
keywords = [
    "novel-writing", "autonomous-agents", "llm", "creative-writing",
    "self-publishing", "kdp", "epub", "multi-agent", "openclaw"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: General",
]
dependencies = [
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "rich>=13.7",
    "click>=8.1",
    "jinja2>=3.1",
    "ebooklib>=0.18",
    "reportlab>=4.1",
    "pypdf>=4.0",
    "beautifulsoup4>=4.12",
    "lxml>=5.1",
    "pydantic>=2.6",
    "tiktoken>=0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.3",
    "mypy>=1.8",
]

[project.scripts]
novelclaw = "autonovelclaw.cli:main"

[tool.setuptools.packages.find]
include = ["autonovelclaw*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
